Added PodSpec

+++ DKAudioPlayer/DKAudioPlayer.m
@@ -1,7 +1,27 @@
 //
 //  DKAudioPlayer.m
 //
-//  Created by Dennis Kutlubaev on 17.01.14.
+//  Created by Dennis Kutlubaev on 27.02.14.
+//  This code is distributed under the terms and conditions of the MIT license.
+//  Copyright (c) 2014 Dennis Kutlubaev (kutlubaev.denis@gmail.com)
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
 //
 
 #import "DKAudioPlayer.h"
@@ -179,9 +199,9 @@ blue:((float)(rgbValue & 0xFF))/255.0 alpha:opacity]
 {
     long currentPlaybackTime = _audioPlayer.currentTime;
     
-    int currentHours = (currentPlaybackTime / 3600);
-    int currentMinutes = ((currentPlaybackTime / 60) - currentHours*60);
-    int currentSeconds = (currentPlaybackTime % 60);
+    int currentHours = (int)(currentPlaybackTime / 3600);
+    int currentMinutes = (int)((currentPlaybackTime / 60) - currentHours*60);
+    int currentSeconds = (int)(currentPlaybackTime % 60);
     NSString *currentTimeString = (currentHours > 0) ? [NSString stringWithFormat:@"%i:%02d:%02d", currentHours, currentMinutes, currentSeconds] : [NSString stringWithFormat:@"%02d:%02d", currentMinutes, currentSeconds];
     
     NSString *string = [NSString stringWithFormat:@"%@ / %@", currentTimeString, self.durationString];
